requirements.txt
Django>=4.2
djangorestframework
psycopg2-binary
celery[redis]
pandas
python-decouple
dj-database-url

